+++ tests/unit/test_context.py
@@ -271,33 +271,3 @@ class Test_default_cache_policy:
 
         key = key_module.Key("ThisKind", 0)
         assert context_module._default_cache_policy(key) is False
-
-
-class TestCache:
-    @staticmethod
-    def test_get_and_validate_valid():
-        cache = context_module._Cache()
-        test_entity = mock.Mock(_key="test")
-        cache["test"] = test_entity
-        assert cache.get_and_validate("test") is test_entity
-
-    @staticmethod
-    def test_get_and_validate_invalid():
-        cache = context_module._Cache()
-        test_entity = mock.Mock(_key="test")
-        cache["test"] = test_entity
-        test_entity._key = "changed_key"
-        with pytest.raises(KeyError):
-            cache.get_and_validate("test")
-
-    @staticmethod
-    def test_get_and_validate_none():
-        cache = context_module._Cache()
-        cache["test"] = None
-        assert cache.get_and_validate("test") is None
-
-    @staticmethod
-    def test_get_and_validate_miss():
-        cache = context_module._Cache()
-        with pytest.raises(KeyError):
-            cache.get_and_validate("nonexistent_key")
